Concerns #48 - Move common parts to a common file
