+++ src/app/components/reports/chart-report/chart-report.component.ts
@@ -12,9 +12,10 @@ import {
   Reporter,
   ReportedData} from 'src/app/model/core/reporting';
 import { ChartReportConfigurationComponent } from './chart-report-configuration.component';
-import { Ng2ChartReport, Ng2ChartConfiguration, Ng2ChartReportFactory, Ng2ChartAnnotation } from 'src/app/model/reports/ng2-chart.report';
+import { Ng2ChartReport, Ng2ChartReportFactory } from 'src/app/model/reports/ng2-chart.report';
 import { StringUtils } from 'src/app/model/utils/string-utils';
 import { ChartReportAnnotationComponent } from './chart-report-annotation.component';
+import { ChartConfiguration, ChartAnnotation } from 'src/app/model/reports/chart-report';
 
 @Component({
   selector: 'chart-report',
@@ -64,12 +65,12 @@ export class ChartReportComponent implements AfterViewInit, Report {
   }
 
   ngAfterViewInit() {
-    let configurations: Ng2ChartConfiguration[] = [];
+    let configurations: ChartConfiguration[] = [];
     this.configurationComponents.forEach(configurationComponent => {
       configurations.push(configurationComponent.asNg2ChartConfiguration());
     });
 
-    let annotations: Ng2ChartAnnotation[] = [];
+    let annotations: ChartAnnotation[] = [];
     this.annotationComponents.forEach(annotationComponent => {
       annotations.push(annotationComponent.asNg2ChartAnnotation());
     });
